feat: training modal
